+++ docs/changelog.md
@@ -1,6 +1,9 @@
 # Latest 
 - Add github todo-to-issue
 - Mass issue creation via todo-to-issue
+- Partial implementation of deploy-to-github-pages script (hard because windows doesn't mix with bash)
+- Github pages build for Version v3.0.0
+- Github pages attempt in typescript
 
 # Version 3.0.0: Github release version 
 - Create base document